requirements.txt
flask
flask_cors
flask_restx 
flask-sqlalchemy
flask_marshmallow
python-dotenv